+++ assembly/index.ts
@@ -1,7 +1,7 @@
-import {caller, methodNumber} from "./sdk";
-import {usrForbidden, usrUnhandledMsg} from "./sdk";
-import {ActorID, NO_DATA_BLOCK_ID} from "./sdk";
-import {State} from "./sdk/utils/state/json";
+import {caller, methodNumber} from "fvm-as-sdk/assembly";
+import {usrForbidden, usrUnhandledMsg} from "fvm-as-sdk/assembly";
+import {ActorID, NO_DATA_BLOCK_ID} from "fvm-as-sdk/assembly";
+import {saveState} from "./state";
 
 
 export function invoke(_: u32): u32 {
@@ -11,9 +11,9 @@ export function invoke(_: u32): u32 {
     case 1:
       constructor()
       break
-    case 2:
+    /*case 2:
       say_hello()
-      break
+      break*/
     default:
       usrUnhandledMsg()
   }
@@ -23,21 +23,22 @@ export function invoke(_: u32): u32 {
 
 function constructor(): void {
   // This constant should be part of the SDK.
-  const INIT_ACTOR_ADDR: ActorID = 1;
+  const INIT_ACTOR_ADDR: ActorID = 10000;
 
-  if ( caller() != INIT_ACTOR_ADDR ) usrForbidden()
+  //if ( caller() != INIT_ACTOR_ADDR ) usrForbidden()
 
-  const state = new State(0)
-  state.save()
+  /*const state = new State(0)
+  state.save()*/
+  saveState(0)
 
   return;
 }
 
 // Not working
 function say_hello(): Uint8Array {
-  const state = State.load();
+  /*const state = State.load();
   state.count += 1;
-  state.save();
+  state.save();*/
 
   const ret = new Uint8Array(2); // "A"
   ret[0] = 97;
