--- assembly/sdk/index.ts
@@ -1,2 +1,0 @@
-export * from "./wrappers"
-export * from "./env"
