--- assembly/sdk/wrappers/index.ts
@@ -1,3 +1,0 @@
-export * from "./ipld"
-export * from "./message"
-export * from "./errors"
