+++ testing/Cargo.toml
@@ -6,10 +6,10 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-fvm_ipld_blockstore = { version = "0.1.0", path = "../../ref-fvm/ipld/blockstore" }
+fvm_ipld_blockstore = { version = "0.1.1", path = "../../ref-fvm/ipld/blockstore" }
 fvm_integration_tests = { version = "0.1.0", path = "../../ref-fvm/testing/integration" }
-fvm_shared = { version = "0.6.1", path = "../../ref-fvm/shared" }
-fvm = { version = "0.7.1", path = "../../ref-fvm/fvm", default-features = false }
+fvm_shared = { version = "0.7.0", path = "../../ref-fvm/shared" }
+fvm = { version = "1.0.0-rc.1", path = "../../ref-fvm/fvm", default-features = false }
 #fvm_ipld_blockstore = { version = "0.1.0", git = "https://github.com/filecoin-project/ref-fvm", branch = "experimental/fvm-m2" }
 #fvm_integration_tests = { version = "0.1.0", git = "https://github.com/filecoin-project/ref-fvm", branch = "experimental/fvm-m2" }
 #fvm_shared = { version = "0.6.1", git = "https://github.com/filecoin-project/ref-fvm", branch = "experimental/fvm-m2" }
